+++ src/Reed/Classes/Article/ArticleSearchBar.swift
@@ -20,7 +20,6 @@ protocol SearchBarDelegate: NSObjectProtocol {
     optional func searchBar(_ searchBar: ArticleSearchBar, textDidChange: String)
 }
 
-@IBDesignable
 final class ArticleSearchBar: UIStackView {
     var searchField: UISearchTextField!
     var nextButton: UIButton!