+++ src/Reed/Classes/UIComponents/InteractiveLabel.swift
@@ -8,7 +8,6 @@
 
 import UIKit
 
-@IBDesignable
 final class InteractiveLabel: UILabel, @preconcurrency UIEditMenuInteractionDelegate {
     override init(frame: CGRect) {
         super.init(frame: frame)