+++ src/Reed/Classes/Web/HTTPConditionalGetInfo.swift
@@ -26,12 +26,6 @@ nonisolated struct HTTPConditionalGetInfo: Codable, Equatable {
         self.init(lastModified: lastModified, etag: etag)
     }
 
-    init?(headers: [AnyHashable: Any]) {
-        let lastModified = headers[HTTPResponseHeader.lastModified] as? String
-        let etag = headers[HTTPResponseHeader.etag] as? String
-        self.init(lastModified: lastModified, etag: etag)
-    }
-
     func addRequestHeadersToURLRequest(_ urlRequest: inout URLRequest) {
         // Bug seen in the wild: lastModified with last possible 32-bit date, which is in 2038. Ignore those.
         // TODO: drop this check in late 2037.