+++ src/Reed/Classes/Web/HTTPRequestHeader.swift
@@ -10,8 +10,6 @@ import Foundation
 
 nonisolated enum HTTPRequestHeader {
     static let userAgent = "User-Agent"
-    static let authorization = "Authorization"
-    static let contentType = "Content-Type"
 
     // Conditional GET
 
@@ -20,11 +18,6 @@ nonisolated enum HTTPRequestHeader {
 }
 
 nonisolated struct HTTPResponseHeader: Sendable {
-    static let contentType = "Content-Type"
-    static let location = "Location"
-    static let link = "Links"
-    static let date = "Date"
-
     // Conditional GET
     static let lastModified = "Last-Modified"
     static let etag = "Etag"